+++ .gitignore
@@ -6,7 +6,6 @@ tsconfig.tsbuildinfo
 target/
 !migrations/*.sql
 gui/vite.config.ts.timestamp-*
-token_list.json
 
 # extension artifacts
 extension/*.zip
