+++ bin/iron/Cargo.toml
@@ -43,9 +43,6 @@ auto-launch = "0.5.0"
 
 [build-dependencies]
 tauri-build = { version = "1.2", features = [] }
-reqwest = { workspace = true }
-serde = { workspace = true }
-serde_json = { workspace = true }
 iron-token-list = { workspace = true }
 
 [features]
