+++ crates/token-list/Cargo.toml
@@ -8,10 +8,10 @@ repository.workspace = true
 exclude.workspace = true
 authors.workspace = true
 
-[dependencies]
+[build-dependencies]
 chrono = { version = "0.4.31", features = ["serde"] }
-reqwest = { workspace = true }
+reqwest = { workspace = true, features = ["blocking", "json"] }
 serde = { workspace = true }
 serde_json = { workspace = true }
 tokio = { workspace = true }
-thiserror = { workspace = true }
+anyhow = "1.0.79"
